backend/api/filters.py
from django_filters.rest_framework import FilterSet, filters
from recipes.models import Ingredient, Recipe, Tag
from rest_framework.filters import SearchFilter


class IngredientFilter(SearchFilter):
    search_param = 'name'

    class Meta:
        model = Ingredient
        fields = ('name',)


class RecipeFilter(FilterSet):
    tags = filters.ModelMultipleChoiceFilter(
        field_name='tags__slug',
        to_field_name='slug',
        queryset=Tag.objects.all(),
    )
    is_favorite = filters.NumberFilter(method='filter_is_favorite')
    is_in_cart = filters.NumberFilter(method='filter_is_in_cart')

    class Meta:
        model = Recipe
        fields = ('tags', 'author', 'is_favorite', 'is_in_cart',)

    def filter_is_favorite(self, queryset, name, value):
        if value and self.request.user.is_authenticated:
            return queryset.filter(favorites__user=self.request.user)
        else:
            return queryset

    def filter_is_in_cart(self, queryset, name, value):
        if value and self.request.user.is_authenticated:
            return queryset.filter(shopping_list__user=self.request.user)
        else:
            return queryset
