requirements.txt
asgiref==3.2.10
requests==2.26.0
django==2.2.16
django-filter==2.4.0
djangorestframework==3.12.4
gunicorn==20.0.4
psycopg2-binary==2.8.6
PyJWT==2.1.0
pytz==2020.1
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
sqlparse==0.3.1
python-dotenv
djoser==2.1.0
drf_extra_fields
drf-yasg
django-colorfield